util/log: move the crdb-v1 test code to a separate file
Release note: None
